+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - make possible to decide which properties the data browser will filter on
 - add "datalayers" query string parameter to override shown datalayers on map load
 - add edit fieldset for changing marker latlng by hand
+- moved from Leaflet.Draw to Leaflet.Editable
 
 
 ## 0.6.x